refactor: major refactor to sql confogurable stack and config write
